Fix a bug where non zero positions are incorrect for bitarray
Since bitarray 0.9.0, the `numpy.where(bitarray)` call does not return
the correct non zero positions for bits. Instead, it returns the index
of non zero bytes. For example, for `bitarray("0000000001000000")`, the
function call returns non zero positions `[1]`, instead of `[9]`, since
bitarray release 0.9.0.

This change fixes this issue by iterating over the bitarray and returning
the list of indices at which the element is `True`.